2.0.2-RELEASE, correctly delete identifiers when deleting fhir links for patient (testfix)

+++ root/api/src/main/java/org/patientview/api/service/impl/UserServiceImpl.java
@@ -566,13 +566,15 @@ public class UserServiceImpl extends AbstractServiceImpl<UserServiceImpl> implem
         Set<Long> fhirLinkIdentifierIds = new HashSet<>();
         
         User user = userRepository.findOne(userId);
-        for (FhirLink fhirLink : user.getFhirLinks()) {
-            fhirLinkIdentifierIds.add(fhirLink.getIdentifier().getId());
-            fhirLinkRepository.delete(fhirLink.getId());
-        }
-        
-        for(Long id : fhirLinkIdentifierIds) {
-            identifierRepository.delete(id);
+        if (user.getFhirLinks() != null) {
+            for (FhirLink fhirLink : user.getFhirLinks()) {
+                fhirLinkIdentifierIds.add(fhirLink.getIdentifier().getId());
+                fhirLinkRepository.delete(fhirLink.getId());
+            }
+
+            for (Long id : fhirLinkIdentifierIds) {
+                identifierRepository.delete(id);
+            }
         }
 
         user.setFhirLinks(new HashSet<FhirLink>());
